+++ Public/Engine/Start-Workflow.ps1
@@ -4,13 +4,13 @@ function Start-Workflow {
         Run Alteryx workflow
 
         .DESCRIPTION
-        Run an Alteryx workflow by command-line
+        Run an Alteryx workflow from command-line
 
         .NOTES
         File name:      Start-Workflow.ps1
         Author:         Florian Carrier
         Creation date:  2021-09-15
-        Last modified:  2021-09-15
+        Last modified:  2021-11-20
     #>
     [CmdletBinding ()]
     Param (
@@ -32,45 +32,33 @@ function Start-Workflow {
         [System.IO.FileInfo]
         $AppValues,
         [Parameter (
-            Position    = 3,
-            Mandatory   = $false,
-            HelpMessage = "Version of the engine to use"
-        )]
-        [ValidateSet (
-            "1",
-            "2",
-            "Original",
-            "AMP"
-        )]
-        [System.String]
-        $EngineVersion,
-        [Parameter (
             Position    = 4,
             Mandatory   = $false,
             HelpMessage = "Path to the Alteryx engine command line utility"
         )]
         [ValidateNotNullOrEmpty ()]
-        [Alias ("Path")]
+        [Alias ("AlteryxEngine")]
         [System.IO.FileInfo]
-        $EnginePath
+        $AlteryxEngine,
+        [Parameter (
+            HelpMessage = "Switch to force the use of the AMP engine"
+        )]
+        [Switch]
+        $AMP
     )
     Begin {
         # Get global preference vrariables
         Get-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
         # Engine path
-        if (-Not $PSBoundParameters.ContainsKey("EnginePath")) {
-            $EnginePath = Get-Utility -Utility "Engine"
+        if (-Not $PSBoundParameters.ContainsKey("AlteryxEngine")) {
+            $AlteryxEngine = Get-Utility -Utility "Engine"
         }
     }
     Process {
         # Parameters
         $Parameters = ([System.String]::Concat('"', $Workflow, '"', " ", $AppValues)).Trim()
-        # Call engine
-        if ($PSBoundParameters.ContainsKey("EngineVersion")) {
-            $Output = Invoke-Engine -Path $EnginePath -Parameters $Parameters -EngineVersion $EngineVersion
-        } else {
-            $Output = Invoke-Engine -Path $EnginePath -Parameters $Parameters
-        }
+        # Run workflow
+        $Output = Invoke-Engine -Path $AlteryxEngine -Parameters $Parameters -AMP:$AMP
         return $Output
     }
 }