+++ Public/General/Set-ServerSecret.ps1
@@ -10,7 +10,7 @@ function Set-ServerSecret {
         File name:      Set-ServerSecret.ps1
         Author:         Florian Carrier
         Creation date:  2021-08-27
-        Last modified:  2021-09-20
+        Last modified:  2021-10-27
     #>
     [CmdletBinding ()]
     Param (
@@ -30,7 +30,7 @@ function Set-ServerSecret {
         )]
         [ValidateNotNullOrEmpty ()]
         [Alias ("ServicePath")]
-        [String]
+        [System.IO.FileInfo]
         $Path
     )
     Begin {