+++ Public/License/Add-License.ps1
@@ -10,7 +10,7 @@ function Add-License {
         File name:      Add-License.ps1
         Author:         Florian Carrier
         Creation date:  2021-06-05
-        Last modified:  2021-09-20
+        Last modified:  2021-10-27
 
         .LINK
         https://www.powershellgallery.com/packages/PSAYX
@@ -47,7 +47,7 @@ function Add-License {
             HelpMessage = "Path to Alteryx licensing utility"
         )]
         [ValidateNotNullOrEmpty ()]
-        [String]
+        [System.IO.FileInfo]
         $Path,
         [Parameter (
             HelpMessage = "Switch to suppress log generation"