+++ Public/License/Reset-LicenseServer.ps1
@@ -10,7 +10,7 @@ function Reset-LicenseServer {
         File name:      Reset-LicenseServer.ps1
         Author:         Florian Carrier
         Creation date:  2021-06-09
-        Last modified:  2021-09-20
+        Last modified:  2021-10-27
         Comment:        **Untested**
 
         .LINK
@@ -28,7 +28,7 @@ function Reset-LicenseServer {
             HelpMessage = "Path to Alteryx licensing utility"
         )]
         [ValidateNotNullOrEmpty ()]
-        [String]
+        [System.IO.FileInfo]
         $Path,
         [Parameter (
             HelpMessage = "Switch to suppress log generation"